* working on tests for scalajs
