+++ core_tests/jvm/src/main/java/com/paulbutcher/test/JavaGenericInterface.java
