+++ core_tests/jvm/src/main/java/com/paulbutcher/test/PolymorphicJavaInterface.java
