+++ core_tests/jvm/src/test/scala/com.paulbutcher.test/mock/JavaMocksTest.scala
@@ -110,4 +110,6 @@ class JavaMocksTest extends IsolatedSpec {
     m.overloadedGeneric("one") shouldBe "first"
     m.overloadedGeneric(2) shouldBe "second"
   }
+
+  override def newInstance = new JavaMocksTest
 }
