+++ core_tests/shared/src/test/scala/com/paulbutcher/test/mock/MethodsWithDefaultParamsTest.scala
@@ -83,4 +83,6 @@ class MethodsWithDefaultParamsTest extends IsolatedSpec {
     m.withAllDefaultParams()
     m.withAllDefaultParams("other", CaseClass(99))
   }
+
+  override def newInstance = new MethodsWithDefaultParamsTest
 }
