+++ core_tests/shared/src/test/scala/com/paulbutcher/test/mock/OverloadedMethodsTest.scala
@@ -116,4 +116,5 @@ class OverloadedMethodsTest extends IsolatedSpec {
     m.print("foo")
   }
 
+  override def newInstance = new OverloadedMethodsTest
 }
